move ChatGPTApp out of page to fix build issue(#28)
